Rhapsody plugin with class diagram
