--- RhapsodyPlugin/Project_rpy/Default.sbsx
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<RhapsodyArchive>
-	<MagicNumber></MagicNumber>
-	<CODE-PAGE>1252</CODE-PAGE>
-	<version>110.0.1</version>
-	<lang>C++</lang>
-	<BuildNo>202411191641</BuildNo>
-	<RMMMinimumClientVersion>100.0</RMMMinimumClientVersion>
-	
-	<RHAPSODY-MODEL>
-	<ISubsystem type="e">
-			<_id type="a">GUID 4b86ab5f-e7de-448a-803f-b023a5406bb7</_id>
-			<_myState type="a">8192</_myState>
-			<_name type="a">Default</_name>
-			<_modifiedTimeWeak type="a">4.23.2025::10:32:56</_modifiedTimeWeak>
-			<_modifiedTime type="a">4.23.2025::10:32:56</_modifiedTime>
-			<_lastID type="a">3</_lastID>
-			<weakCGTime type="a">4.23.2025::10:32:56</weakCGTime>
-			<strongCGTime type="a">1.2.1990::0:0:0</strongCGTime>
-			<_defaultComposite type="s">GUID 7894f864-fc6f-46ab-8040-beba5975192a</_defaultComposite>
-			<_eventsBaseID type="a">-1</_eventsBaseID>
-			<_configurationRelatedTime type="a">4.23.2025::10:32:56</_configurationRelatedTime>
-			<AggregatesList type="e">
-				<value>GUID 7894f864-fc6f-46ab-8040-beba5975192a</value>
-			</AggregatesList>
-		</ISubsystem>
-		<IClass type="e">
-			<_id type="a">GUID 7894f864-fc6f-46ab-8040-beba5975192a</_id>
-			<_myState type="a">40960</_myState>
-			<_name type="a">TopLevel</_name>
-			<_modifiedTimeWeak type="a">1.2.1990::0:0:0</_modifiedTimeWeak>
-			<_modifiedTime type="a">1.2.1990::0:0:0</_modifiedTime>
-			<weakCGTime type="a">1.2.1990::0:0:0</weakCGTime>
-			<strongCGTime type="a">1.2.1990::0:0:0</strongCGTime>
-			<_itsStateChart type="r">
-				<IHandle type="e">
-				</IHandle>
-			</_itsStateChart>
-			<_classModifier type="a">Unspecified</_classModifier>
-		</IClass>
-		
-	</RHAPSODY-MODEL>
-	<OslcLinks type="c">
-		<IRPYRawContainer type="e">
-		</IRPYRawContainer>
-	</OslcLinks>
-
-<INCLUDES-RHAPSODY-INDEX>false</INCLUDES-RHAPSODY-INDEX>
-</RhapsodyArchive>
-
