--- RhapsodyPlugin/Project_rpy/DefaultComponent.cmpx
@@ -1,131 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<RhapsodyArchive>
-	<MagicNumber></MagicNumber>
-	<CODE-PAGE>1252</CODE-PAGE>
-	<version>110.0.1</version>
-	<lang>C++</lang>
-	<BuildNo>202408271326</BuildNo>
-	<RMMMinimumClientVersion>100.0</RMMMinimumClientVersion>
-	
-	<RHAPSODY-MODEL>
-	<IComponent type="e">
-			<_id type="a">GUID 25ecca31-85fc-4ccd-a50d-e628e1e7429a</_id>
-			<_myState type="a">139264</_myState>
-			<_name type="a">DefaultComponent</_name>
-			<_modifiedTimeWeak type="a">1.2.1990::0:0:0</_modifiedTimeWeak>
-			<_modifiedTime type="a">3.27.2025::4:41:2</_modifiedTime>
-			<m_buildType type="a">Executable</m_buildType>
-			<m_configActive type="r">
-				<ICodeGenConfigInfoHandle type="e">
-					<_hm2Class type="a">ICodeGenConfigInfo</_hm2Class>
-					<_hid type="a">GUID c9503b81-0c97-4d81-8a9f-4422684033fe</_hid>
-				</ICodeGenConfigInfoHandle>
-			</m_configActive>
-			<HandlesInMe type="c">
-				<IRPYRawContainer type="e">
-					
-						<IHandle type="e">
-							<_hm2Class type="a">ISubsystem</_hm2Class>
-							<_hfilename type="a">.\Default.sbsx</_hfilename>
-							<_hname type="a">Default</_hname>
-							<_hid type="a">GUID 4b86ab5f-e7de-448a-803f-b023a5406bb7</_hid>
-						</IHandle>
-					
-					
-						<IHandle type="e">
-							<_hm2Class type="a">IProfile</_hm2Class>
-							<_hfilename type="a">$OMROOT\Settings\UMLPerspectives\UMLPerspectives.sbs</_hfilename>
-							<_hname type="a">UMLPerspectives</_hname>
-							<_hid type="a">GUID d24d9192-bc4f-48a2-a4e1-fdaebbdfa6b6</_hid>
-						</IHandle>
-					
-				</IRPYRawContainer>
-			</HandlesInMe>
-			<VariationPoints type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</VariationPoints>
-			<SelectedVariants type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</SelectedVariants>
-			<weakCGTime type="a">1.2.1990::0:0:0</weakCGTime>
-			<strongCGTime type="a">1.2.1990::0:0:0</strongCGTime>
-			<AggregatesList type="e">
-				<value>GUID 01759690-fca7-4d12-8815-ecb7832f6b3e</value>
-				<value>GUID c9503b81-0c97-4d81-8a9f-4422684033fe</value>
-			</AggregatesList>
-		</IComponent>
-		<IFolder type="e">
-			<_id type="a">GUID 01759690-fca7-4d12-8815-ecb7832f6b3e</_id>
-			<_myState type="a">73728</_myState>
-			<_name type="a">Files</_name>
-			<_modifiedTimeWeak type="a">3.27.2025::4:41:2</_modifiedTimeWeak>
-			<_modifiedTime type="a">3.27.2025::4:41:2</_modifiedTime>
-			<Elements type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</Elements>
-			<Files type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</Files>
-		</IFolder>
-		<ICodeGenConfigInfo type="e">
-			<_id type="a">GUID c9503b81-0c97-4d81-8a9f-4422684033fe</_id>
-			<_myState type="a">8192</_myState>
-			<_name type="a">DefaultConfig</_name>
-			<_modifiedTimeWeak type="a">1.2.1990::0:0:0</_modifiedTimeWeak>
-			<_modifiedTime type="a">3.27.2025::4:41:2</_modifiedTime>
-			<_scopeType type="a">Explicit</_scopeType>
-			<_instrumentation type="a">None</_instrumentation>
-			<_timeModel type="a">RealTime</_timeModel>
-			<m_generateActors type="a">1</m_generateActors>
-			<_statechartImplementation type="a">Flat</_statechartImplementation>
-			<_checksList type="a">0 ,</_checksList>
-			<_checksList_Default type="a">1</_checksList_Default>
-			<ScopeElements type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</ScopeElements>
-			<InitialInstances type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</InitialInstances>
-			<weakCGTime type="a">1.2.1990::0:0:0</weakCGTime>
-			<strongCGTime type="a">1.2.1990::0:0:0</strongCGTime>
-			<AnimScopeElements type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</AnimScopeElements>
-			<m_allInAnimScope type="a">1</m_allInAnimScope>
-			<m_generateUsecases type="a">0</m_generateUsecases>
-			<AggregatesList type="e">
-				<value>GUID e0dc8aa2-b7d3-4e1e-bfb6-9ac460f201f0</value>
-			</AggregatesList>
-		</ICodeGenConfigInfo>
-		<IFolder type="e">
-			<_id type="a">GUID e0dc8aa2-b7d3-4e1e-bfb6-9ac460f201f0</_id>
-			<_myState type="a">8192</_myState>
-			<_name type="a">DefaultConfig</_name>
-			<_modifiedTimeWeak type="a">1.2.1990::0:0:0</_modifiedTimeWeak>
-			<_modifiedTime type="a">3.27.2025::4:41:2</_modifiedTime>
-			<Elements type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</Elements>
-			<Files type="c">
-				<IRPYRawContainer type="e">
-				</IRPYRawContainer>
-			</Files>
-		</IFolder>
-		
-	</RHAPSODY-MODEL>
-	<OslcLinks type="c">
-		<IRPYRawContainer type="e">
-		</IRPYRawContainer>
-	</OslcLinks>
-
-<INCLUDES-RHAPSODY-INDEX>false</INCLUDES-RHAPSODY-INDEX>
-</RhapsodyArchive>
-
